Intellithesis/requirements.txt
Django>=5.0.7
djangorestframework>=3.14.0
groq>=0.4.0
PyMuPDF>=1.23.0
python-docx>=0.8.11
psycopg2-binary>=2.9.0 